se genera el modulo de gestión de proveedores, se agregó al app.blade.php abajo de Ordenes de Compra. También se modificó el tema de registro de Insumos, para calcular el Precio de venta, pendiente generar un modulo para gestionar estos valores de %ganancia y %iva en dado caso llegue a cambiar en el futuro

+++ database/migrations/2025_10_01_052229_create_insumos_table.php
@@ -10,8 +10,8 @@ return new class extends Migration {
             $table->id();
             $table->string('nombre', 50);
             $table->decimal('costo', 10, 2)->nullable();
-            $table->decimal('stock', 10, 2)->nulleable();
-            $table->decimal('stock_minimo', 10, 2);
+            $table->integer('stock')->nulleable();
+            $table->integer('stock_minimo');
             $table->string('descripcion', 200);
             $table->unsignedBigInteger('type_insumo_id');
             $table->decimal('precio', 10, 2)->nullable();
